backend/requirements.txt
pandas
scikit-learn
joblib
numpy
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
python-dotenv
requests